% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_switch.R
\name{app_data_switch}
\alias{app_data_switch}
\title{App data switch (function data switch)}
\usage{
app_data_switch(data, app)
}
\arguments{
\item{data}{name of data}

\item{app}{logical, app data or not}
}
\value{
Lookup table
}
\description{
App data switch (function data switch)
}
\examples{
app_data_switch(data = "RiskPairData", app = TRUE)
}
