% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extdata}
\alias{get_extdata}
\title{Import data files from inst/extdata/ folder}
\usage{
get_extdata(file_name)
}
\arguments{
\item{file_name}{name of file}
}
\value{
data file
}
\description{
Intended to be used inside pakcage folder
}
\examples{
# meant to be used for package development!
# get_extdata("RiskPairData.csv")
# get_extdata("AppRiskPairData.csv")
}
