% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_importFrom}
\alias{get_importFrom}
\title{Get roxygen2 tags (\verb{@importFrom})}
\usage{
get_importFrom(pkg, fun = NULL)
}
\arguments{
\item{pkg}{package (must be installed)}

\item{fun}{string of function names}
}
\value{
tags for \code{roxygen2}
}
\description{
Get roxygen2 tags (\verb{@importFrom})
}
\examples{
require(bs4Dash)
get_importFrom(pkg = 'bs4Dash', fun = 'valueBox')
}
