% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_import_data.R
\name{import_flat_file}
\alias{import_flat_file}
\title{Import plain text data file}
\usage{
import_flat_file(path)
}
\arguments{
\item{path}{path to file}
}
\value{
imported flat file
}
\description{
Import plain text data file
}
