% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_extdata_paths}
\alias{list_extdata_paths}
\title{Find inst/extdata/ paths}
\usage{
list_extdata_paths()
}
\value{
list of absolute file paths for \code{inst/extdata}
}
\description{
Find inst/extdata/ paths
}
\examples{
list_extdata_paths()
}
