% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_tables.R
\name{seg_bgm_ref_vars}
\alias{seg_bgm_ref_vars}
\title{SEG BMR/REF Variables}
\usage{
seg_bgm_ref_vars(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}
}
\value{
seg BGM/REF included and excluded variables
}
\description{
SEG BMR/REF Variables
}
\examples{
seg_bgm_ref_vars(
data = system.file("extdata", "VanderbiltComplete.csv",
        package = "segtools"),
is_path = TRUE)
}
