% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_binom_table.R
\name{seg_binom_table}
\alias{seg_binom_table}
\title{Create compliant pairs table}
\usage{
seg_binom_table(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{Logical, dataset or path to dataset}
}
\value{
Binomial test table
}
\description{
Create compliant pairs table
}
\examples{
seg_binom_table(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)
}
