% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_binom_tbl.R
\name{seg_binom_tbl}
\alias{seg_binom_tbl}
\title{Create compliant pairs table}
\usage{
seg_binom_tbl(risk_vars)
}
\arguments{
\item{risk_vars}{output from \code{seg_risk_vars()}}
}
\value{
Binomial test table
}
\description{
Create compliant pairs table
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
risk_cols_tbl <- seg_risk_vars(df = test_data)
seg_binom_tbl(risk_cols_tbl)
}
