% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_graph.R
\name{seg_graph}
\alias{seg_graph}
\title{ggplot2 function for SEG grid}
\usage{
seg_graph(
  data,
  alpha_var = 0.6,
  size_var = 2.3,
  color_var = "#000000",
  fill_var = "#FFFFFF"
)
}
\arguments{
\item{data}{Dataset containing only \code{BGM} and \code{REF} columns}

\item{alpha_var}{alpha setting}

\item{size_var}{size setting}

\item{color_var}{color setting}

\item{fill_var}{fill setting}
}
\value{
SEG grid graph
}
\description{
ggplot2 function for SEG grid
}
