% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_grid.R
\name{seg_grid}
\alias{seg_grid}
\title{ggplot2 function for SEG grid}
\usage{
seg_grid(data, is_path, alpha_var, size_var, color_var)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}

\item{alpha_var}{alpha setting}

\item{size_var}{size setting}

\item{color_var}{color setting}
}
\value{
SEG grid graph
}
\description{
ggplot2 function for SEG grid
}
