% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_vars.R
\name{seg_iso_cols}
\alias{seg_iso_cols}
\title{Title}
\usage{
seg_iso_cols(risk_cat_cols)
}
\arguments{
\item{risk_cat_cols}{output from \code{seg_risk_cat_cols()}}
}
\value{
ISO range variable table
}
\description{
Title
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
risk_cat_tbl <- seg_risk_cat_cols(test_data)
seg_iso_cols(risk_cat_cols = risk_cat_tbl)
}
