% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_iso_range_table.R
\name{seg_iso_range_table}
\alias{seg_iso_range_table}
\title{Create ISO range table}
\usage{
seg_iso_range_table(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{Logical, dataset or path to dataset}
}
\value{
ISO range (and category) tibble
}
\description{
Create ISO range table
}
\examples{
seg_iso_range_table(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)
}
