% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_cols.R
\name{seg_iso_vars}
\alias{seg_iso_vars}
\title{Title}
\usage{
seg_iso_vars(risk_cat_vars)
}
\arguments{
\item{risk_cat_vars}{output from \code{seg_risk_cat_vars()}}
}
\value{
ISO range variable table
}
\description{
Title
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
risk_cat_tbl <- seg_risk_cat_vars(test_data)
seg_iso_vars(risk_cat_vars = risk_cat_tbl)
}
