% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_mard_table.R
\name{seg_mard_table}
\alias{seg_mard_table}
\title{SEG MARD table}
\usage{
seg_mard_table(data, is_path = FALSE)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{Logical, dataset or path to dataset}
}
\value{
MARD table
}
\description{
Mean absolute relative difference (MARD) table. See
    [article](https://journals.sagepub.com/doi/10.1177/1932296816662047)
    for more information,
}
\examples{
seg_mard_table(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)
}
