% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_modba.R
\name{seg_modba}
\alias{seg_modba}
\title{The SEG modified Bland-Altman plot}
\usage{
seg_modba(data)
}
\arguments{
\item{data}{Dataset containing only \code{BGM} and \code{REF} columns}

\item{is_path}{logical, dataset or path to dataset}
}
\value{
modified Bland-Altman plot
}
\description{
The SEG modified Bland-Altman plot
}
\examples{
require(vroom)
vanderbilt_complete <- vroom::vroom(
 file =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"), delim = ",")
seg_modba(data = vanderbilt_complete, is_path = FALSE)
}
