% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_modba_graph.R
\name{seg_modba_graph}
\alias{seg_modba_graph}
\title{The SEG modified Bland-Altman plot}
\usage{
seg_modba_graph(data, text_size = 16)
}
\arguments{
\item{data}{Dataset containing only \code{BGM} and \code{REF} columns}

\item{text_size}{base size passed to \code{theme_seg()}}
}
\value{
modified Bland-Altman plot
}
\description{
The SEG modified Bland-Altman plot
}
\examples{
require(vroom)
vanderbilt_complete <- vroom::vroom(
 file =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"), delim = ",")
seg_modba_graph(data = vanderbilt_complete)
}
