% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_pair_type_tbl.R
\name{seg_pair_type_tbl}
\alias{seg_pair_type_tbl}
\title{SEG pairs type table}
\usage{
seg_pair_type_tbl(df)
}
\arguments{
\item{df}{data with \code{BGM} and \code{REF} columns}
}
\value{
pair type columns
}
\description{
SEG pairs type table
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
seg_pair_type_tbl(test_data)
}
