% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.R
\name{seg_pairs_table}
\alias{seg_pairs_table}
\title{Create Pair Type Table}
\usage{
seg_pairs_table(data, is_path = FALSE)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}
}
\value{
pair type table
}
\description{
Create Pair Type Table
}
\examples{
seg_pairs_table(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)

}
