% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_cols.R
\name{seg_risk_cat_vars}
\alias{seg_risk_cat_vars}
\title{SEG risk category variables}
\usage{
seg_risk_cat_vars(df)
}
\arguments{
\item{df}{data with \code{BGM} and \code{REF} columns}
}
\value{
seg risk category variable table
}
\description{
SEG risk category variables
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
seg_risk_cat_vars(test_data)
}
