% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_cats.R, R/seg_tables.R
\name{seg_risk_cats}
\alias{seg_risk_cats}
\title{SEG Risk Category Variables}
\usage{
seg_risk_cats(data, is_path)

seg_risk_cats(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}
}
\value{
seg risk category variables

seg risk category variables
}
\description{
SEG Risk Category Variables

SEG Risk Category Variables
}
\examples{
seg_risk_cats(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
  is_path = TRUE)
seg_risk_cats(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
  is_path = TRUE)
}
