% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_grade_table.R
\name{seg_risk_grade_table}
\alias{seg_risk_grade_table}
\title{Create risk grade table}
\usage{
seg_risk_grade_table(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{Logical, dataset or path to dataset}
}
\value{
risk grade tibble
}
\description{
Create risk grade table
}
\examples{
seg_risk_grade_table(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)
}
