% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_level_table.R
\name{seg_risk_level_table}
\alias{seg_risk_level_table}
\title{Create risk level (and category) table}
\usage{
seg_risk_level_table(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{Logical, dataset or path to dataset}
}
\value{
Risk level (and category) tibble
}
\description{
Create risk level (and category) table
}
\examples{
seg_risk_level_table(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)
}
