% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_pairs.R, R/seg_tables.R
\name{seg_risk_pairs}
\alias{seg_risk_pairs}
\title{SEG Risk Pair Variables}
\usage{
seg_risk_pairs(data, is_path)

seg_risk_pairs(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}

\item{app_data}{logical, include data with `App` prefix}
}
\value{
seg risk pair variables

seg risk pair variables
}
\description{
SEG Risk Pair Variables

SEG Risk Pair Variables
}
\examples{
seg_risk_pairs(data = system.file("extdata", "VanderbiltComplete.csv",
       package = "segtools"), is_path = TRUE)
seg_risk_pairs(data = system.file("extdata", "VanderbiltComplete.csv",
       package = "segtools"), is_path = TRUE)
}
