% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_vars.R
\name{seg_risk_vars}
\alias{seg_risk_vars}
\title{SEG risk vars}
\usage{
seg_risk_vars(data, is_path, app_data)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}
}
\value{
SEG risk variables
}
\description{
SEG risk vars
}
\examples{
seg_risk_vars(data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
            is_path = TRUE)
}
