% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_seg.R
\name{theme_seg}
\alias{theme_seg}
\title{ggplot2 theme (for SEG grid)}
\usage{
theme_seg(
  base_size = 14,
  base_family = "Ubuntu",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}

\item{base_line_size}{base graph line size}

\item{base_rect_size}{base graph rectangle size}
}
\value{
ggplot2 theme
}
\description{
ggplot2 theme (for SEG grid)
}
