% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_flat_file.R
\name{import_flat_file}
\alias{import_flat_file}
\title{Import plain text data file}
\usage{
import_flat_file(path)
}
\arguments{
\item{path}{path to file}
}
\value{
imported flat file
}
\description{
Import plain text data file
}
\examples{
test_data <- import_flat_file(system.file("extdata", "VanderbiltComplete.csv",
                              package = "segtools"))
utils::str(test_data)
}
