% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_mard_tbl.R
\name{seg_mard_tbl}
\alias{seg_mard_tbl}
\title{SEG MARD table}
\usage{
seg_mard_tbl(risk_vars)
}
\arguments{
\item{risk_vars}{output from \code{seg_risk_vars()}}
}
\value{
MARD table
}
\description{
SEG MARD table
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
risk_cols_tbl <- seg_risk_vars(df = test_data)
seg_mard_tbl(risk_cols_tbl)
}
