% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_grade_tbl.R
\name{seg_risk_grade_tbl}
\alias{seg_risk_grade_tbl}
\title{SEG risk grade table}
\usage{
seg_risk_grade_tbl(risk_cols)
}
\arguments{
\item{risk_cols}{output from \code{seg_risk_cols()}}
}
\value{
risk grade table
}
\description{
SEG risk grade table
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
risk_cols_tbl <- seg_risk_cols(df = test_data)
seg_risk_grade_tbl(risk_cols_tbl)
}
