% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_risk_vars.R
\name{seg_risk_vars}
\alias{seg_risk_vars}
\title{SEG risk columns (wrapper function)}
\usage{
seg_risk_vars(df)
}
\arguments{
\item{df}{data with \code{BGM} and \code{REF} columns}
}
\value{
SEG risk columns
}
\description{
SEG risk columns (wrapper function)
}
\examples{
test_data <- vroom::vroom(
                system.file("extdata", "VanderbiltComplete.csv",
                package = "segtools"), delim = ",")
seg_risk_vars(df = test_data)
}
