% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.R
\name{seg_risk_vars}
\alias{seg_risk_vars}
\title{SEG Risk Variables}
\usage{
seg_risk_vars(data, is_path)
}
\arguments{
\item{data}{Dataset containing only `BGM` and `REF` columns}

\item{is_path}{logical, dataset or path to dataset}
}
\value{
seg risk table variables
}
\description{
SEG Risk Variables
}
\examples{
seg_risk_vars(
 data =
   system.file("extdata", "VanderbiltComplete.csv",
               package = "segtools"),
is_path = TRUE)

}
